+++ src/com/cop5536/Ride.java
@@ -2,19 +2,20 @@ package com.cop5536;
 
 public class Ride {
 
-	long rideNumber;
+	int rideNumber;
 	int rideCost;
 	int tripDuration;
 
 	int pos;
 
-	public Ride(long rideNumber, int rideCost, int tripDuration) {
+	public Ride(int rideNumber, int rideCost, int tripDuration) {
 		this.rideNumber = rideNumber;
 		this.rideCost = rideCost;
 		this.tripDuration = tripDuration;
 		this.pos = 0;
 	}
-	public Ride(long rideNumber) {
+
+	public Ride(int rideNumber) {
 		this.rideNumber = rideNumber;
 	}
